+++ system/src/apps/LevelUpSD.mjs
@@ -57,9 +57,14 @@ export default class LevelUpSD extends FormApplication {
 			event => this._onReRollHP(event)
 		);
 
+		html.find("[data-action='view-boon-table']").click(
+			event => this._viewBoonTable(event)
+		);
+
 		html.find("[data-action='view-talent-table']").click(
 			event => this._viewTalentTable(event)
 		);
+
 		html.find("[data-action='open-spellbook']").click(
 			event => this._openSpellBook(event)
 		);
@@ -70,6 +75,10 @@ export default class LevelUpSD extends FormApplication {
 
 		html.find("[data-action='delete-spell']").click(
 			event => this._onDeleteSpell(event)
+		);
+
+		html.find("[data-action='roll-boon']").click(
+			event => this._onRollBoon(event)
 		);
 
 		html.find("[data-action='roll-talent']").click(
@@ -93,9 +102,26 @@ export default class LevelUpSD extends FormApplication {
 			this.data.class = await fromUuid(this.data.actor.system.class);
 			this.data.talentTable = await fromUuid(this.data.class.system.classTalentTable);
 			this.data.currentLevel = this.data.actor.system.level.value;
-			this.data.targetLevel =  this.data.currentLevel +1;
+			this.data.targetLevel =  this.data.currentLevel + 1;
 			this.data.talentGained = (this.data.targetLevel % 2 !== 0);
 			this.data.totalSpellsToChoose = 0;
+
+			this.data.patron = await fromUuid(this.data.actor.system.patron);
+
+			this.data.boonTable = this.data.patron
+				? await fromUuid(this.data.patron.system.boonTable)
+				: undefined;
+
+			this.data.startingBoons = 0;
+			const classData = this.data.class.system;
+			this.data.canRollBoons = classData.patron.required;
+
+			const needsStartingBoons = classData.patron.required
+				&& classData.patron.startingBoons > 0;
+
+			if (this.data.targetLevel === 1 && needsStartingBoons) {
+				this.data.startingBoons = classData.patron.startingBoons;
+			}
 
 			if (await this.data.actor.isSpellCaster()) {
 				this.data.spellcastingClass =
@@ -129,6 +155,7 @@ export default class LevelUpSD extends FormApplication {
 
 			}
 		}
+
 		return this.data;
 	}
 
@@ -152,6 +179,10 @@ export default class LevelUpSD extends FormApplication {
 		}
 	}
 
+	async _viewBoonTable() {
+		this.data.boonTable.sheet.render(true);
+	}
+
 	async _viewTalentTable() {
 		this.data.talentTable.sheet.render(true);
 	}
@@ -160,13 +191,17 @@ export default class LevelUpSD extends FormApplication {
 		this.data.actor.openSpellBook();
 	}
 
-	async _onRollHP() {
+	async _onRollHP({isReroll = false}) {
 		const data = {
 			rollType: "hp",
 			actor: this.data.actor,
 		};
 		let options = {};
-		options.title = game.i18n.localize("SHADOWDARK.dialog.hp_roll.title");
+
+		options.title = isReroll
+			? game.i18n.localize("SHADOWDARK.dialog.hp_re_roll.title")
+			: game.i18n.localize("SHADOWDARK.dialog.hp_roll.title");
+
 		options.flavor = options.title;
 		options.chatCardTemplate = "systems/shadowdark/templates/chat/roll-hp.hbs";
 		options.fastForward = true;
@@ -186,10 +221,28 @@ export default class LevelUpSD extends FormApplication {
 		Dialog.confirm({
 			title: "Re-Roll HP",
 			content: "Are you sure you want to re-roll hit points?",
-			yes: () => this._onRollHP(),
+			yes: () => this._onRollHP({isReroll: true}),
 			no: () => null,
 			defaultYes: false,
 		  });
+	}
+
+	async _onRollBoon() {
+		if (!this.data.boonTable) {
+			return ui.notifications.warn(
+				game.i18n.localize("SHADOWDARK.apps.level-up.errors.missing_boon_table")
+			);
+		}
+
+		await this.data.boonTable.draw();
+		ui.sidebar.activateTab("chat");
+
+		if (this.data.targetLevel > 1) {
+			this.data.rolls.talent = true;
+		}
+		this.data.rolls.boon = true;
+
+		this.render();
 	}
 
 	async _onRollTalent() {
@@ -208,7 +261,11 @@ export default class LevelUpSD extends FormApplication {
 			}
 		}
 
+		if (this.data.targetLevel > 1) {
+			this.data.rolls.boon = true;
+		}
 		this.data.rolls.talent = true;
+
 		this.render();
 	}